♻️ Refactor: stream, interaction, draw 분리
